Fix binary workflow (#40)
* Make .exe extension optional

Uses extended globbing to enable the use of `@` which enables us to have an optional .exe extension. This is done instead of `{,.exe}` which expands to `nimlangserver nimlangserver.exe` and then fails since it cant find the `.exe` file

* Fix formatting in `on` section

Needed to do this so I can test my changes via `workflow_dispatch`

* Fix `nim-version` input

Before it was passing to the input `nimversion` which doesn't exist. This makes it pass to the correct input `nim-version` and actual gives it a value since `nimversion` doesn't exist in the matrix.

Technically this can be removed since stable is the default value, but this means that it is easier to change version in future