Server shutdown in case the client process has exited unexpectedly. (#64)
Monitor the client process ID, received in the initialize request and shutdown
the server in case the client process is found to be dead. This fixes the
remaining nimlangserver processes after an IDE crash, or in case of broken IDE
and/or plugin not sending the proper shutdown and exit commands to the server.