Gets the nimsuggest path from `nimble` when `nimble` has `nimdir`
This depends on https://github.com/nim-lang/nimble/pull/1221
It also needs to be tested in Win.

While developing this found an issue with the project discovery, it uses `nim dump` but it doesnt seem to end when Nim is a local dep, it takes too long. Will try to speed it up in the nimble side of things